requirements.txt
torch==1.2.0
torchvision==0.4.0
runway-python==0.4.0